Commercial hedging pressure (CHP) factor

Provided with futures contract front price and term structure aggregated position data from Bloomberg retrieved with pullit, construct commercial hedging pressure factor. The futures commercial hedging pressure factor is based on the well-known hedging pressure-based theory (Anderson and Danthine 1983; Chang 1985; Cootner 1960; Dusak 1973; Hicks 1939; Hirshleifer 1988, 1989; Kolb 1992; Keynes 1930) which postulates that futures prices for a given commodity are inversely related to the extent that commercial hedgers are short or long and the mimicking portfolio here aims at capturing the impact of hedging pressure as a systemic factor (Basu and Miffre 2013) .

CHP_factor(
  price_data,
  CHP_data,
  update_frequency = "month",
  return_frequency = "day",
  ranking_period = 6L,
  long_threshold = 0.5,
  short_threshold = 0.5,
  weighted = F
)

# S4 method for FuturesTS,FuturesCFTC
CHP_factor(
  price_data,
  CHP_data,
  update_frequency = "month",
  return_frequency = "day",
  ranking_period = 6L,
  long_threshold = 0.5,
  short_threshold = 0.5,
  weighted = F
)

Arguments

price_data
an S4 object of class FuturesTS. FuturesTS objects are returned by the BBG_futures_TS function in the pullit package.

CHP_data
an S4 object of class FuturesCFTC. FuturesCFTC objects are returned by the BBG_futures_CFTC function in the pullit package.

update_frequency
a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', 'semester', 'quarter', 'month' or 'week'. Defaults to 'month'.

return_frequency
a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', 'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.

ranking_period
a scalar integer vector. Specifies number of periods in term of update_frequency looking backward for average CHP calculation. Defaults to 1 where sort is done on last observation only.

long_threshold
a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.

short_threshold
a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.

weighted
a scalar logical vector. If 'TRUE' adjusts portoflio weights with respect to pressure, else equal weights are used. Defaults to 'FALSE'.

Value

An S4 object of class CHPFactor.

References

Anderson RW, Danthine J (1983). “Hedger diversity in futures markets.” The Economic Journal, 93(370), 370--389. ISSN 0013-0133, doi: 10.2307/2232798 .

Basu D, Miffre J (2013). “Capturing the risk premium of commodity futures: the role of hedging pressure.” Journal of Banking & Finance, 37(7), 2652--2664. doi: 10.1016/j.jbankfin.2013.02.031 .

Chang EC (1985). “Returns to speculators and the theory of normal backwardation.” The Journal of Finance, 40(1), 193--208. doi: 10.1111/j.1540-6261.1985.tb04944.x .

Cootner PH (1960). “Returns to speculators: Telser versus Keynes.” Journal of Political Economy, 68(4), 396--404. doi: 10.1086/258347 .

Dusak K (1973). “Futures trading and investor returns: an investigation of commodity market risk premiums.” Journal of Political economy, 81(6), 1387--1406. doi: 10.1086/260133 .

Hicks JR (1939). Value and capital. Oxford University Press, Cambridge.

Hirshleifer D (1988). “Risk, futures pricing, and the organization of production in commodity markets.” Journal of Political Economy, 96(6), 1206--1220. ISSN 0022-3808, doi: 10.1086/261584 .

Hirshleifer D (1989). “Determinants of hedging and risk premia in commodity futures markets.” Journal of Financial and Quantitative Analysis, 24(3), 313--331. ISSN 1756-6916, 0022-1090, doi: 10.2307/2330814 .

Hirshleifer D (1990). “Hedging pressure and futures price movements in a general equilibrium model.” Econometrica, 58(2), 411--428. ISSN 0012-9682, doi: 10.2307/2938209 .

Keynes JM (1930). Treatise on money. Macmillan, London.

Kolb RW (1992). “Is normal backwardation normal?” Journal of Futures Markets, 12(1), 75--91.