Fama-McBeth cross-sectional regression

Given two dataframes of assets returns and factors returns respectively, performs a two-pass cross-sectional regression following Fama and MacBeth (1973) . The method attempts to test the relationship between assets average return and systematic factor risk.

famamcbeth(assets_returns, factor_returns, means)

Arguments

assets_returns
a dataframe with dates in the first column and assets returns in subsequent columns.

factor_returns
a dataframe with dates in the first column and factor returns in the second.

means
a scalar logical vector. If TRUE, the second step of the cross-sectional regression is implemented with asset mean returns. In this setting, tests of significance for the correpsonding lambda coefficients are performed with autocorrelation and heteroskedasticity robust standard errors, following Newey and West (1987) . If FALSE, asset returns are regressed on the corresponding betas for each observation date. The corresponding lambda coefficients are averaged and significance is tested using classic mean t-tests.

Value

An S4 object of class FamaMcBeth.

References

Fama EF, MacBeth JD (1973). “Risk, return, and equilibrium: empirical tests.” Journal of political economy, 81(3), 607--636. doi: 10.1086/260061 , https://doi.org/10.1086/260061.

Newey WK, West KD (1987). “A simple, positive semi-definite, heteroskedasticity and autocorrelationconsistent covariance matrix.” Econometrica, 55, 703--708.