Commercial hedging pressure (CHP) factor

Provided with futures contract front price and term structure aggregated position data from Bloomberg retrieved with pullit, construct commercial hedging pressure factor.

CHP_factor(price_data, CHP_data, update_frequency, return_frequency,
  ranking_period, long_threshold, short_threshold, geometric)

# S4 method for FuturesTS,FuturesCFTC
CHP_factor(price_data, CHP_data,
  update_frequency = "month", return_frequency = "day",
  ranking_period = 6L, long_threshold = 0.5, short_threshold = 0.5,
  geometric = TRUE)

Arguments

price_data
an S4 object of class FuturesTS. FuturesTS objects are returned by the BBG_futures_TS function in the pullit package.

CHP_data
an S4 object of class FuturesCFTC. FuturesCFTC objects are returned by the BBG_futures_CFTC function in the pullit package.

update_frequency
a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', 'semester', 'quarter', 'month' or 'week'. Defaults to 'month'.

return_frequency
a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', 'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.

ranking_period
a scalar integer vector. Specifies number of periods in term of update_frequency looking backward for average CHP calculation. Defaults to 0 where sort is done on last observation only.

long_threshold
a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.

short_threshold
a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.

geometric
a scalar logical vector. If TRUE geometric returns are returned, else arithmetic. Default: TRUE.

Value

An S4 object of class CHPFactor.